python: implement buffer protocol for IBuffer
This implements the CPython buffer protocol [1] for the
Windows.Storage.Streams.IBuffer interface. This allows Python
to access the buffer memory directly via memoryview(), the
struct module and so on.

This is also a significant performance improvement for applications
that just need to copy the data to/from the buffer.

[1]: https://docs.python.org/3/c-api/buffer.html