fix duplicate TypeVar definitions
This fixes mypy complaining about duplicate TypeVar definitions. For
delegates, we had to change from typing.Callable to typing.Protocol
to be able to move the TypeVars inside of the class definition since
inheriting from typing.Callable is not allowed. This also has the
advantage of being able to see the input parameter names.

Fixes: https://github.com/pywinrt/pywinrt/issues/12