use sequence mixins for vector types
This makes types that implement IVector and IVectorView inherit
collections.abc.MutableSequence and Sequence respectively. A `insert`
attribute is added that is an alias for the InsertAt method to provide
the required abstract method implementation. Slicing and proper
exceptions are still not implemented since there isn't a runtime type
that implements this that is easy to test.