pywinrt: move test for is_overloaded
The is_overloaded parameter was not used in most cases, so remove it
from the enumerate_methods() callback. Instead we introduce a new helper
the re-enumerates the methods. This new method catches a case that was
missed before, that is when an interface contains a method with the
same name as a required interface.