move top-level package to system package
This makes the top level package a namespace package since it no longer
has an __init__.py. This will enable us to split the package into
multiple distributions in the future.

Also, having a `system` package makes the naming similar to .NET as
a place where we can import fundamental types like `Array`.