pywinrt: call PyObject_GC_UnTrack in tp_dealloc
Since we initialize types with PyType_FromSpecWithBases(), it is
possible that a base type may enable Py_TPFLAGS_HAVE_GC so we need to
always check and make sure we call PyObject_GC_UnTrack() if that is the
case.

Using the debug build of Python 3.11, we would get:

    .gc:0: ResourceWarning: Object of type _winrt.Array is not untracked before destruction

when running tests because the Array type inherits from
abc.MutableSequence but we weren't calling the untrack function.