no longer inherit from collections.abc
In Python 3.12, it is no longer possible for C types to inherit
from abc.ABC since it has a custom tp_new. So we must find an
alternate solution to implement collections.abc protocols.

This attempts to do that by still using the methods from
collection.abc but mixing in the methods manually and registering
the type.

Idea from: https://github.com/python/cpython/issues/103968#issuecomment-1589928055