projection: drop get_unknown() hacks
Now that we are calling proper constructors, we no longer need the hack
to get the instance's IUnknown* to be able to call the as<T>() method.

This also cleans up the pinterface wrapper to properly inherit from the
regular object wrapper so we don't get any suprises if we are operating
on the base system.Object type.