projection: relax runtime type checking
For composed runtime types, we may need to pass objects to an API that
is expecting a "subclass". We haven't impemented the type inheritence
for this in Python but we can at least allow the APIs to accept objects
if we can query the required interface on the WinRT IUnknown object.

Issue: https://github.com/pywinrt/pywinrt/issues/25