projection: release GIL in Application.start()
The Application.start() methods are long running since they pump events
for the application. So we need to release the GIL to allow other Python
threads to run.

Since these are the only known methods that need to release the GIL, we
just hack in a special test for them.