PyWinRT: add --component-dlls option
This adds a new option to PyWinRT.exe for generating a Python projection
for user components when the component DLLs will be shipped in the
Python package.

In this case ApiInformation::Is<Member>Present checks are not needed
because the matching version of the component DLLs will be present.
And calling this would fail anyway for unpackaged apps.

Also, CppWinRT already has a mechanism for loading the component DLLs,
but it won't be able to find them if we don't specify the path to them.
So we add a new helper function to the winrt-runtime package to make
this trivial for component packages.

Issue: https://github.com/pywinrt/pywinrt/issues/24