projection: Updated to Windows App SDK 1.6.240829007
We can drop -Wno-missing-field-initializers since the bug has been fixed.

We also need to download the Microsoft.Web.WebView2 nuget package
separately. This was apparently included in Windows App SDK 1.5 but
is no longer included in 1.6.

Eventually, we should create a separate source package for this but
for now, we will include it in the Windows App SDK as before.