projection: fix buffer protocol type hints for array parameters
PyWinRT supports passing any buffer protocol object to WinRT APIs that
that take a WinRT array. We can use the PEP 688 collections.abc.Buffer
type to indicate this. We also keep the existing Array[T] annotations
so that we know what the element type is.

Since PEP 688 didn't address the issue of readable vs writable buffers,
we create two type aliases to indicate if the array should already
contain data (ReadableBuffer) or if the array will be filled with data
(WritableBuffer).

Fixes: https://github.com/pywinrt/pywinrt/issues/81