projection: fix WebView2 package
The package is actually a component package. It can't have the runtime
API support checks. They always return false and fail.  Also, there is
a runtime .dll file that needs to be included, otherwise the WebView2
control will not work. A sample app is added to test that the WebView2
is working.