projection: use Overload attribute for method names
There were some APIs that were not callable because they had the same
number of parameters as another method. In many cases, the WinRT
metadata has an Windows.Foundation.Metadata.Overload attribute on these
methods to disambiguate them. This changes the code generator to use
this attribute as the method name when present. This allows the missing
methods to be exposed, but at the expense of renaming quite a few other
methods.