projection: split types with circular dependencies into separate binary WriterExtensions
In some cases, composable classes have circular dependencies where a
type in namespace A inherits from a type in namespace B, which in turn
inherits from a type in namespace A. This will cause problems when we
try to implement proper inheritance in Python as it will create circular
imports. To work around this, create separate binary extensions so that
we can import things in stages.