projection: fix interface type hints
Since system.Object is the real base class, it needs be first in the
list of base classes. However, when interface types were listed after
this, we were getting MRO errors from mypy.

To work around this, create a new interface type that doesn't inherit
from system.Object, and use that for the interface type hints for
inheritance and for input parameters. Output parameters will still use
the runtime object interface type since it returns an actually runtime
class wrapper object specifically for the interface.

We don't use typing.Protocol because this is not duck typing.