projection: implement toggle reference
For Python subclasses of composable WinRT classes, we need to ensure
that both the Python object and the WinRT object are kept alive as long
there is a reference to one or the other.

In order to do this, both the Python object and the WinRT object need to
hold a reference to each other. But this creates a reference cycle, so
we need to break the cycle when the reference count of the WinRT object
drops to 1. To do this, we add a toggle_reference method that hooks into
AddRef and Release and gets called back when the reference count changes
to or from 1.

When the py_obj_ref object is constructed, we start with a weak
reference to the Python object. If the WinRT object ref count increases,
we change the weak reference to a strong reference. If the WinRT object
ref count decreases to 1, we change the strong reference back to a weak
reference. If there are no more references to the Python object, the
object is destroyed.