projection: handle round-trip of user-defined subclasses
Before now, it didn't matter if new instances of wrapper classes were
created each time we marshal a WinRT object to Python, because there is
no state stored in the Python object.

Now, since we can create user-defined subclasses of WinRT classes, we
need to ensure that we return the same Python object each time so that
we don't lose any of the state that is implemented in Python.

This is done by adding a new interface for getting the Python object
pointer and implementing it. Then when we check for this interface first
before creating a new Python object.