projection: use IInspectable for composable types
Since composable types can be used as base classes, the WinRT object
pointer may not be of the same type as the Python wrapper type. So we
need to use IInspectable instead of the exact type.

This also revealed some bugs in sequence, map and closable types where
we weren't calling QueryInterface on the WinRT object pointer before
using it to call a method.