projection: implement calling protected methods
Finish implementation of calling protected methods.

Calling overridable methods is tricky because the WinRT implementation
calls the method override of the most derived class, even parent classes.
So we have to add a bit of a hack to get the inner pointer of Python
subclass to be able to avoid recursive calls to the Python override
of the method.