pybase: protect python_implements against GIL race
The implementation of abi_enter() and abi_exit() to manage the GIL state
would fail if more than one thread was calling methods of the same
object at the same time due to the shared state variable.

Instead, use a local variable in each function to store the GIL state
to protect against this.