projection: release GIL when calling all WinRT methods
Although most WinRT methods don't need to release the GIL because they
aren't waiting on other threads that may call back into Python and
should be short running functions, there is no way to know for sure
which methods actually need it. So, to be safe, release the GIL for all
WinRT methods. For example, updating the dict returned by the
FileSavePicker.file_type_choices property will cause a deadlock if the
GIL is not released because it handles it on a background thread which
may call back into Python, e.g. if a Python list was passed as one of
the values.

Unfortunately, this change will have a performance impact on all WinRT
methods, but it is necessary to avoid deadlocks.