projection: implement __instancecheck__ for winrt interface checking
This does a better job of checking if Object._as(IInterface) will succeed.

Since this method has to be implemented in a metaclass and due to
metaclass inheritance requirements, all interfaces have to have a
metaclass now. Runtime classes also get this same metaclass via the
system.Object base class.