projection: fix isinstance for dynamic interfaces
Add missing feature from cppsrint to also allow extending GetIids when
extra interfaces are added using query_interface_tearoff. This allows
runtime introspection of which interfaces are implemented without
calling QueryInterface.

Also fix a bug in the existing implementation for composeable classs
that was causing the wrong ponter to be returned from NonDelagatingGetIids.