projection: make struct attributes readonly
Since structs are copy-by-value, it is easy to try to modify a copy of
a struct rather than the struct itself when referencing the struct as
an attribute.

For example:

    window = Window()
    window.size.x = 100

Does not set the size of the window, but rather sets x on a copy of the
size struct returned by getting the window.size attribute.