projection: reorder constexpr specializations in headers
Some headers also include temlates for genric types that make use of
constexpr specializations of some types defined in pybase.h. In order
to avoid a compiler error, these specializations need to be defined
before they are used. In order to do this, move the definitions before
the #include of the py.Namespace.h header files.