projection: allow tuple as input arg for projected struct
It can be quite verbose to have to write the type of each struct when
calling a function or setting a property.  Allow a plain tuple to be
passed in these cases instead. Should be more efficient too.

Note: mypy 1.16 will be required for type checking to work correctly
on properties.