projection: move non-Windows SDK libs to different root packages
Both WinUI 2 and Win UI 3 use the same namespaces, so we need to use
the top-level package a a qualifier to differentiate between the two.

This also allows an extra level of organization for the packages so that
we don't have everything at the same level in the projection folder.