projection: implement __await__ in Python and fix up
Move implementation of __await__ to Python code. This makes it easier to
see what is going on and makes for less repeated generated code.

This also includes a few fixes. The cancellation of the Future that wraps
an async operation now is propagated. Also, if get_results() fails, e.g.
due to a missing import, the exception will be propagated via the
Future rather than silently failing and blocking forever.

Fixes: https://github.com/pywinrt/pywinrt/issues/112