projection: make wrap_async internal and change implementation
Instead of using a Future, implement wrap_async using a coroutine.
This simplifies the implementation and should make it robust against
never completing. It will now also properly wait for cancellation of
the underlying operation instead of immediately canceling the Future
while the underlying operation is still running.

The function is moved to a new runtime._internals module to make it
clear that it is not part of the public API. There isn't really a need
for it to be public now that it is implemented as a coroutine.

Now asyncio.create_task() and asyncio.ensure_future() will work as expected
rather than wrapping the Future in another Future or Task.