projection: pybase: handle conversion of winrt::param::hstring
There is a new API that has a progress handler that takes a string
parameter. This is the first time we've seen a string parameter in
a delegate. It uses winrt::param::hstring instead of winrt::hstring,
so we need to be able to handle that to avoid a compilation error.